+++ lib/Immediate/REPL.cpp
@@ -31,8 +31,12 @@
 #include "llvm/ExecutionEngine/MCJIT.h"
 #include "llvm/IR/Module.h"
 #include "llvm/IR/Constants.h"
+#include "llvm/IR/DiagnosticInfo.h"
+#include "llvm/IR/DiagnosticPrinter.h"
 #include "llvm/Transforms/Utils/Cloning.h"
 #include "llvm/ADT/StringSet.h"
+#include "llvm/IRReader/IRReader.h"
+#include "llvm/Linker/Linker.h"
 #include "llvm/Support/ConvertUTF.h"
 #include "llvm/Support/PrettyStackTrace.h"
 #include "llvm/Support/Process.h"
@@ -224,6 +228,34 @@ public:
   
   void print(llvm::raw_ostream &out) const override;
 };
+} // end anonymous namespace
+
+namespace {
+static void linkerDiagnosticHandlerNoCtx(const llvm::DiagnosticInfo &DI) {
+  if (DI.getSeverity() != llvm::DS_Error)
+    return;
+
+  std::string MsgStorage;
+  {
+    llvm::raw_string_ostream Stream(MsgStorage);
+    llvm::DiagnosticPrinterRawOStream DP(Stream);
+    DI.print(DP);
+  }
+  llvm::errs() << "Error linking swift modules\n";
+  llvm::errs() << MsgStorage << "\n";
+}
+
+
+
+static void linkerDiagnosticHandler(const llvm::DiagnosticInfo &DI,
+                                    void *Context) {
+  // This assert self documents our precondition that Context is always
+  // nullptr. It seems that parts of LLVM are using the flexibility of having a
+  // context. We don't really care about this.
+  assert(Context == nullptr && "We assume Context is always a nullptr");
+
+  return linkerDiagnosticHandlerNoCtx(DI);
+}
 } // end anonymous namespace
 
 /// EditLine wrapper that implements the user interface behavior for reading
@@ -847,6 +879,44 @@ private:
         }
       }
     }
+  }
+
+  bool linkLLVMModules(llvm::Module *Module,
+                       std::unique_ptr<llvm::Module> &&ModuleToLink) {
+    // EGREGIOUS HACKS AHEAD
+    //
+    // 1) LLVMContext does not support robust notions of identity rdar://61895075
+    // 2) llvm::Linker does not support modules allocated in separate contexts
+    //    rdar://61894890
+    // 3) Cloning modules across contexts is a known deficiency in LLVM
+    //    rdar://61896692
+    // 4) Round-tripping through the IR printer and parser is not guaranteed to
+    //    result in the same IR. It is heavily tested and implied that this is
+    //    the case, but LLVM makes no formal guarantees.
+    //
+    // So, work around 1) and do a naive pointer comparison to see if we
+    // allocated the module we're about to link in a separate context.
+    if (&Module->getContext() != &ModuleToLink->getContext()) {
+      // If so, work around 2) by round-tripping the IR and parsing back it
+      // into the original context. With module in-hand, we have successfully
+      // worked around 3).
+      llvm::SmallString<1024> scratch;
+      llvm::raw_svector_ostream PrintBuffer(scratch);
+      ModuleToLink->print(PrintBuffer, nullptr);
+      auto Buffer = llvm::MemoryBuffer::getMemBufferCopy(PrintBuffer.str());
+      llvm::SMDiagnostic Err;
+      // Finally, hope that 4) doesn't come back to bite us in the long run.
+      ModuleToLink = llvm::parseIR(Buffer->getMemBufferRef(), Err,
+                                   Module->getContext());
+    }
+
+    llvm::LLVMContext &Ctx = Module->getContext();
+    auto OldHandler = Ctx.getDiagnosticHandlerCallBack();
+    void *OldDiagnosticContext = Ctx.getDiagnosticContext();
+    Ctx.setDiagnosticHandlerCallBack(linkerDiagnosticHandler, nullptr);
+    bool Failed = llvm::Linker::linkModules(*Module, std::move(ModuleToLink));
+    Ctx.setDiagnosticHandlerCallBack(OldHandler, OldDiagnosticContext);
+    return !Failed;
   }
 
   bool executeSwiftSource(llvm::StringRef Line, const ProcessCmdLine &CmdLine) {
